add license header and comments
