+++ lib/src/main/res/raw/trust.js
@@ -642,11 +642,13 @@ var e=require("./web3/requestmanager"),t=require("./web3/iban"),r=require("./web
 var e=require("./lib/web3");"undefined"!=typeof window&&void 0===window.Web3&&(window.Web3=e),module.exports=e;
 },{"./lib/web3":"OHXT"}],"afne":[function(require,module,exports) {
 "use strict";function _classCallCheck(e,r){if(!(e instanceof r))throw new TypeError("Cannot call a class as a function")}function _defineProperties(e,r){for(var t=0;t<r.length;t++){var n=r[t];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(e,n.key,n)}}function _createClass(e,r,t){return r&&_defineProperties(e.prototype,r),t&&_defineProperties(e,t),e}var RPCServer=function(){function e(r){_classCallCheck(this,e),this.rpcUrl=r}return _createClass(e,[{key:"getBlockNumber",value:function(){return this.call({jsonrpc:"2.0",method:"eth_blockNumber",params:[]}).then(function(e){return e.result})}},{key:"getBlockByNumber",value:function(e){return this.call({jsonrpc:"2.0",method:"eth_getBlockByNumber",params:[e,!1]}).then(function(e){return e.result})}},{key:"getFilterLogs",value:function(e){return this.call({jsonrpc:"2.0",method:"eth_getLogs",params:[e]})}},{key:"call",value:function(e){return fetch(this.rpcUrl,{method:"POST",headers:{Accept:"application/json","Content-Type":"application/json"},body:JSON.stringify(e)}).then(function(e){return e.json()}).then(function(e){if(!e.result&&e.error)throw console.log("<== rpc error",e.error),new Error(e.error.message||"rpc error");return e})}}]),e}();module.exports=RPCServer;
+},{}],"MNFS":[function(require,module,exports) {
+"use strict";function _typeof(t){return(_typeof="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function _classCallCheck(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function _defineProperties(t,e){for(var r=0;r<e.length;r++){var o=e[r];o.enumerable=o.enumerable||!1,o.configurable=!0,"value"in o&&(o.writable=!0),Object.defineProperty(t,o.key,o)}}function _createClass(t,e,r){return e&&_defineProperties(t.prototype,e),r&&_defineProperties(t,r),t}function _possibleConstructorReturn(t,e){return!e||"object"!==_typeof(e)&&"function"!=typeof e?_assertThisInitialized(t):e}function _assertThisInitialized(t){if(void 0===t)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return t}function _inherits(t,e){if("function"!=typeof e&&null!==e)throw new TypeError("Super expression must either be null or a function");t.prototype=Object.create(e&&e.prototype,{constructor:{value:t,writable:!0,configurable:!0}}),e&&_setPrototypeOf(t,e)}function _wrapNativeSuper(t){var e="function"==typeof Map?new Map:void 0;return(_wrapNativeSuper=function(t){if(null===t||!_isNativeFunction(t))return t;if("function"!=typeof t)throw new TypeError("Super expression must either be null or a function");if(void 0!==e){if(e.has(t))return e.get(t);e.set(t,r)}function r(){return _construct(t,arguments,_getPrototypeOf(this).constructor)}return r.prototype=Object.create(t.prototype,{constructor:{value:r,enumerable:!1,writable:!0,configurable:!0}}),_setPrototypeOf(r,t)})(t)}function isNativeReflectConstruct(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],function(){})),!0}catch(t){return!1}}function _construct(t,e,r){return(_construct=isNativeReflectConstruct()?Reflect.construct:function(t,e,r){var o=[null];o.push.apply(o,e);var n=new(Function.bind.apply(t,o));return r&&_setPrototypeOf(n,r.prototype),n}).apply(null,arguments)}function _isNativeFunction(t){return-1!==Function.toString.call(t).indexOf("[native code]")}function _setPrototypeOf(t,e){return(_setPrototypeOf=Object.setPrototypeOf||function(t,e){return t.__proto__=e,t})(t,e)}function _getPrototypeOf(t){return(_getPrototypeOf=Object.setPrototypeOf?Object.getPrototypeOf:function(t){return t.__proto__||Object.getPrototypeOf(t)})(t)}var ProviderRpcError=function(t){function e(t,r){var o;return _classCallCheck(this,e),(o=_possibleConstructorReturn(this,_getPrototypeOf(e).call(this))).code=t,o.message=r,o}return _inherits(e,_wrapNativeSuper(Error)),_createClass(e,[{key:"toString",value:function(){return"".concat(this.message," (").concat(this.code,")")}}]),e}();module.exports=ProviderRpcError;
 },{}],"FOZT":[function(require,module,exports) {
 "use strict";function _toConsumableArray(e){return _arrayWithoutHoles(e)||_iterableToArray(e)||_nonIterableSpread()}function _nonIterableSpread(){throw new TypeError("Invalid attempt to spread non-iterable instance")}function _iterableToArray(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}function _arrayWithoutHoles(e){if(Array.isArray(e)){for(var r=0,n=new Array(e.length);r<e.length;r++)n[r]=e[r];return n}}function _classCallCheck(e,r){if(!(e instanceof r))throw new TypeError("Cannot call a class as a function")}function _defineProperties(e,r){for(var n=0;n<r.length;n++){var t=r[n];t.enumerable=t.enumerable||!1,t.configurable=!0,"value"in t&&(t.writable=!0),Object.defineProperty(e,t.key,t)}}function _createClass(e,r,n){return r&&_defineProperties(e.prototype,r),n&&_defineProperties(e,n),e}var Utils=function(){function e(){_classCallCheck(this,e)}return _createClass(e,null,[{key:"genId",value:function(){return(new Date).getTime()+Math.floor(1e3*Math.random())}},{key:"flatMap",value:function(e,r){var n;return(n=[]).concat.apply(n,_toConsumableArray(e.map(r)))}},{key:"intRange",value:function(e,r){return e>=r?[]:new Array(r-e).fill().map(function(r,n){return n+e})}},{key:"hexToInt",value:function(e){return null==e?e:Number.parseInt(e,16)}},{key:"intToHex",value:function(e){return null==e?e:"0x"+e.toString(16)}}]),e}();module.exports=Utils;
 },{}],"sD6q":[function(require,module,exports) {
 "use strict";var _utils=_interopRequireDefault(require("./utils"));function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}function _classCallCheck(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}function _defineProperties(e,t){for(var i=0;i<t.length;i++){var n=t[i];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(e,n.key,n)}}function _createClass(e,t,i){return t&&_defineProperties(e.prototype,t),i&&_defineProperties(e,i),e}var IdMapping=function(){function e(){_classCallCheck(this,e),this.intIds=new Map}return _createClass(e,[{key:"tryIntifyId",value:function(e){if(e.id){if("number"!=typeof e.id){var t=_utils.default.genId();this.intIds.set(t,e.id),e.id=t}}else e.id=_utils.default.genId()}},{key:"tryRestoreId",value:function(e){var t=this.tryPopId(e.id);t&&(e.id=t)}},{key:"tryPopId",value:function(e){var t=this.intIds.get(e);return t&&this.intIds.delete(e),t}}]),e}();module.exports=IdMapping;
 },{"./utils":"FOZT"}],"Focm":[function(require,module,exports) {
-"use strict";var _web=_interopRequireDefault(require("web3")),_rpc=_interopRequireDefault(require("./rpc")),_utils=_interopRequireDefault(require("./utils")),_id_mapping=_interopRequireDefault(require("./id_mapping")),_events=require("events");function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}function _typeof(e){return(_typeof="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function _classCallCheck(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}function _defineProperties(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function _createClass(e,t,n){return t&&_defineProperties(e.prototype,t),n&&_defineProperties(e,n),e}function _possibleConstructorReturn(e,t){return!t||"object"!==_typeof(t)&&"function"!=typeof t?_assertThisInitialized(e):t}function _assertThisInitialized(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}function _inherits(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&_setPrototypeOf(e,t)}function _wrapNativeSuper(e){var t="function"==typeof Map?new Map:void 0;return(_wrapNativeSuper=function(e){if(null===e||!_isNativeFunction(e))return e;if("function"!=typeof e)throw new TypeError("Super expression must either be null or a function");if(void 0!==t){if(t.has(e))return t.get(e);t.set(e,n)}function n(){return _construct(e,arguments,_getPrototypeOf(this).constructor)}return n.prototype=Object.create(e.prototype,{constructor:{value:n,enumerable:!1,writable:!0,configurable:!0}}),_setPrototypeOf(n,e)})(e)}function isNativeReflectConstruct(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],function(){})),!0}catch(e){return!1}}function _construct(e,t,n){return(_construct=isNativeReflectConstruct()?Reflect.construct:function(e,t,n){var r=[null];r.push.apply(r,t);var s=new(Function.bind.apply(e,r));return n&&_setPrototypeOf(s,n.prototype),s}).apply(null,arguments)}function _isNativeFunction(e){return-1!==Function.toString.call(e).indexOf("[native code]")}function _setPrototypeOf(e,t){return(_setPrototypeOf=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function _getPrototypeOf(e){return(_getPrototypeOf=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}var ProviderRpcError=function(e){function t(e,n){var r;return _classCallCheck(this,t),(r=_possibleConstructorReturn(this,_getPrototypeOf(t).call(this))).code=e,r.message=n,r}return _inherits(t,_wrapNativeSuper(Error)),_createClass(t,[{key:"toString",value:function(){return"".concat(this.message," (").concat(this.code,")")}}]),t}(),TrustWeb3Provider=function(e){function TrustWeb3Provider(e){var t;return _classCallCheck(this,TrustWeb3Provider),(t=_possibleConstructorReturn(this,_getPrototypeOf(TrustWeb3Provider).call(this))).setConfig(e),t.idMapping=new _id_mapping.default,t.callbacks=new Map,t.wrapResults=new Map,t.isTrust=!0,t.isDebug=!1,t._emitConnect(e.chainId),t}return _inherits(TrustWeb3Provider,_events.EventEmitter),_createClass(TrustWeb3Provider,[{key:"isConnected",value:function(){return!0}},{key:"setAddress",value:function(e){this.address=(e||"").toLowerCase(),this.ready=!!e}},{key:"setConfig",value:function(e){this.setAddress(e.address),this.chainId=e.chainId,this.rpc=new _rpc.default(e.rpcUrl)}},{key:"enable",value:function(){return(this||window.ethereum).request({method:"eth_requestAccounts",params:[]})}},{key:"send",value:function(e){var t={jsonrpc:"2.0",id:e.id};switch(e.method){case"eth_accounts":t.result=this.eth_accounts();break;case"eth_coinbase":t.result=this.eth_coinbase();break;case"net_version":t.result=this.net_version();break;case"eth_chainId":t.result=this.eth_chainId();break;default:throw new ProviderRpcError(4200,"Trust does not support calling ".concat(e.method," synchronously without a callback. Please provide a callback parameter to call ").concat(e.method," asynchronously."))}return t}},{key:"sendAsync",value:function(e,t){Array.isArray(e)?Promise.all(e.map(this._request.bind(this))).then(function(e){return t(null,e)}).catch(function(e){return t(e,null)}):this._request(e).then(function(e){return t(null,e)}).catch(function(e){return t(e,null)})}},{key:"request",value:function(e){return(this||window.ethereum)._request(e,!1)}},{key:"_request",value:function(e){var t=this,n=!(arguments.length>1&&void 0!==arguments[1])||arguments[1];return this.idMapping.tryIntifyId(e),this.isDebug&&console.log("==> _request payload ".concat(JSON.stringify(e))),new Promise(function(r,s){switch(e.id||(e.id=_utils.default.genId()),t.callbacks.set(e.id,function(e,t){e?s(e):r(t)}),t.wrapResults.set(e.id,n),e.method){case"eth_accounts":return t.sendResponse(e.id,t.eth_accounts());case"eth_coinbase":return t.sendResponse(e.id,t.eth_coinbase());case"net_version":return t.sendResponse(e.id,t.net_version());case"eth_chainId":return t.sendResponse(e.id,t.eth_chainId());case"eth_sign":return t.eth_sign(e);case"personal_sign":return t.personal_sign(e);case"personal_ecRecover":return t.personal_ecRecover(e);case"eth_signTypedData":case"eth_signTypedData_v3":return t.eth_signTypedData(e);case"eth_sendTransaction":return t.eth_sendTransaction(e);case"eth_requestAccounts":return t.eth_requestAccounts(e);case"eth_newFilter":case"eth_newBlockFilter":case"eth_newPendingTransactionFilter":case"eth_uninstallFilter":case"eth_getFilterChanges":case"eth_getFilterLogs":case"eth_subscribe":throw new ProviderRpcError(4200,"Trust does not support calling ".concat(e.method,". Please use your own solution"));default:return t.callbacks.delete(e.id),t.wrapResults.delete(e.id),t.rpc.call(e).then(function(e){t.isDebug&&console.log("<== rpc response ".concat(JSON.stringify(e))),r(n?e:e.result)}).catch(s)}})}},{key:"_emitConnect",value:function(e){this.emit("connect",{chainId:e})}},{key:"eth_accounts",value:function(){return this.address?[this.address]:[]}},{key:"eth_coinbase",value:function(){return this.address}},{key:"net_version",value:function(){return this.chainId.toString(10)||null}},{key:"eth_chainId",value:function(){return"0x"+this.chainId.toString(16)}},{key:"eth_sign",value:function(e){this.postMessage("signMessage",e.id,{data:e.params[1]})}},{key:"personal_sign",value:function(e){this.postMessage("signPersonalMessage",e.id,{data:e.params[0]})}},{key:"personal_ecRecover",value:function(e){this.postMessage("ecRecover",e.id,{signature:e.params[1],message:e.params[0]})}},{key:"eth_signTypedData",value:function(e){this.postMessage("signTypedMessage",e.id,{data:e.params[1]})}},{key:"eth_sendTransaction",value:function(e){this.postMessage("signTransaction",e.id,e.params[0])}},{key:"eth_requestAccounts",value:function(e){this.postMessage("requestAccounts",e.id,{})}},{key:"postMessage",value:function(e,t,n){this.ready||"requestAccounts"===e?window.webkit.messageHandlers[e].postMessage({name:e,object:n,id:t}):this.sendError(t,new ProviderRpcError(4100,"provider is not ready"))}},{key:"sendResponse",value:function(e,t){var n=this.idMapping.tryPopId(e)||e,r=this.callbacks.get(e),s=this.wrapResults.get(e),o={jsonrpc:"2.0",id:n};"object"===_typeof(t)&&t.jsonrpc&&t.result?o.result=t.result:o.result=t,this.isDebug&&console.log("<== sendResponse id: ".concat(e,", result: ").concat(t,", data: ").concat(JSON.stringify(o))),r&&(r(null,s?o:t),this.callbacks.delete(e))}},{key:"sendError",value:function(e,t){console.log("<== ".concat(e," sendError ").concat(t));var n=this.callbacks.get(e);n&&(n(t instanceof Error?t:new Error(t),null),this.callbacks.delete(e))}}]),TrustWeb3Provider}();window.Trust=TrustWeb3Provider,window.Web3=_web.default;
-},{"web3":"kdoM","./rpc":"afne","./utils":"FOZT","./id_mapping":"sD6q","events":"wIHY"}]},{},["Focm"], null)
+"use strict";var _web=_interopRequireDefault(require("web3")),_rpc=_interopRequireDefault(require("./rpc")),_error=_interopRequireDefault(require("./error")),_utils=_interopRequireDefault(require("./utils")),_id_mapping=_interopRequireDefault(require("./id_mapping")),_events=require("events");function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}function _typeof(e){return(_typeof="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function _classCallCheck(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}function _defineProperties(e,t){for(var n=0;n<t.length;n++){var s=t[n];s.enumerable=s.enumerable||!1,s.configurable=!0,"value"in s&&(s.writable=!0),Object.defineProperty(e,s.key,s)}}function _createClass(e,t,n){return t&&_defineProperties(e.prototype,t),n&&_defineProperties(e,n),e}function _possibleConstructorReturn(e,t){return!t||"object"!==_typeof(t)&&"function"!=typeof t?_assertThisInitialized(e):t}function _assertThisInitialized(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}function _getPrototypeOf(e){return(_getPrototypeOf=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function _inherits(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&_setPrototypeOf(e,t)}function _setPrototypeOf(e,t){return(_setPrototypeOf=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var TrustWeb3Provider=function(e){function TrustWeb3Provider(e){var t;return _classCallCheck(this,TrustWeb3Provider),(t=_possibleConstructorReturn(this,_getPrototypeOf(TrustWeb3Provider).call(this))).setConfig(e),t.idMapping=new _id_mapping.default,t.callbacks=new Map,t.wrapResults=new Map,t.isTrust=!0,t.isDebug=!1,t.emitConnect(e.chainId),t}return _inherits(TrustWeb3Provider,_events.EventEmitter),_createClass(TrustWeb3Provider,[{key:"setAddress",value:function(e){this.address=(e||"").toLowerCase(),this.ready=!!e}},{key:"setConfig",value:function(e){this.setAddress(e.address),this.chainId=e.chainId,this.rpc=new _rpc.default(e.rpcUrl)}},{key:"request",value:function(e){return(this||window.ethereum)._request(e,!1)}},{key:"isConnected",value:function(){return!0}},{key:"enable",value:function(){return(this||window.ethereum).request({method:"eth_requestAccounts",params:[]})}},{key:"send",value:function(e){var t={jsonrpc:"2.0",id:e.id};switch(e.method){case"eth_accounts":t.result=this.eth_accounts();break;case"eth_coinbase":t.result=this.eth_coinbase();break;case"net_version":t.result=this.net_version();break;case"eth_chainId":t.result=this.eth_chainId();break;default:throw new _error.default(4200,"Trust does not support calling ".concat(e.method," synchronously without a callback. Please provide a callback parameter to call ").concat(e.method," asynchronously."))}return t}},{key:"sendAsync",value:function(e,t){Array.isArray(e)?Promise.all(e.map(this._request.bind(this))).then(function(e){return t(null,e)}).catch(function(e){return t(e,null)}):this._request(e).then(function(e){return t(null,e)}).catch(function(e){return t(e,null)})}},{key:"_request",value:function(e){var t=this,n=!(arguments.length>1&&void 0!==arguments[1])||arguments[1];return this.idMapping.tryIntifyId(e),this.isDebug&&console.log("==> _request payload ".concat(JSON.stringify(e))),new Promise(function(s,r){switch(e.id||(e.id=_utils.default.genId()),t.callbacks.set(e.id,function(e,t){e?r(e):s(t)}),t.wrapResults.set(e.id,n),e.method){case"eth_accounts":return t.sendResponse(e.id,t.eth_accounts());case"eth_coinbase":return t.sendResponse(e.id,t.eth_coinbase());case"net_version":return t.sendResponse(e.id,t.net_version());case"eth_chainId":return t.sendResponse(e.id,t.eth_chainId());case"eth_sign":return t.eth_sign(e);case"personal_sign":return t.personal_sign(e);case"personal_ecRecover":return t.personal_ecRecover(e);case"eth_signTypedData":case"eth_signTypedData_v3":return t.eth_signTypedData(e);case"eth_sendTransaction":return t.eth_sendTransaction(e);case"eth_requestAccounts":return t.eth_requestAccounts(e);case"eth_newFilter":case"eth_newBlockFilter":case"eth_newPendingTransactionFilter":case"eth_uninstallFilter":case"eth_getFilterChanges":case"eth_getFilterLogs":case"eth_subscribe":throw new _error.default(4200,"Trust does not support calling ".concat(e.method,". Please use your own solution"));default:return t.callbacks.delete(e.id),t.wrapResults.delete(e.id),t.rpc.call(e).then(function(e){t.isDebug&&console.log("<== rpc response ".concat(JSON.stringify(e))),s(n?e:e.result)}).catch(r)}})}},{key:"emitConnect",value:function(e){this.emit("connect",{chainId:e})}},{key:"eth_accounts",value:function(){return this.address?[this.address]:[]}},{key:"eth_coinbase",value:function(){return this.address}},{key:"net_version",value:function(){return this.chainId.toString(10)||null}},{key:"eth_chainId",value:function(){return"0x"+this.chainId.toString(16)}},{key:"eth_sign",value:function(e){this.postMessage("signMessage",e.id,{data:e.params[1]})}},{key:"personal_sign",value:function(e){this.postMessage("signPersonalMessage",e.id,{data:e.params[0]})}},{key:"personal_ecRecover",value:function(e){this.postMessage("ecRecover",e.id,{signature:e.params[1],message:e.params[0]})}},{key:"eth_signTypedData",value:function(e){this.postMessage("signTypedMessage",e.id,{data:e.params[1]})}},{key:"eth_sendTransaction",value:function(e){this.postMessage("signTransaction",e.id,e.params[0])}},{key:"eth_requestAccounts",value:function(e){this.postMessage("requestAccounts",e.id,{})}},{key:"postMessage",value:function(e,t,n){this.ready||"requestAccounts"===e?window.webkit.messageHandlers[e].postMessage({name:e,object:n,id:t}):this.sendError(t,new _error.default(4100,"provider is not ready"))}},{key:"sendResponse",value:function(e,t){var n=this.idMapping.tryPopId(e)||e,s=this.callbacks.get(e),r=this.wrapResults.get(e),i={jsonrpc:"2.0",id:n};"object"===_typeof(t)&&t.jsonrpc&&t.result?i.result=t.result:i.result=t,this.isDebug&&console.log("<== sendResponse id: ".concat(e,", result: ").concat(t,", data: ").concat(JSON.stringify(i))),s&&(s(null,r?i:t),this.callbacks.delete(e))}},{key:"sendError",value:function(e,t){console.log("<== ".concat(e," sendError ").concat(t));var n=this.callbacks.get(e);n&&(n(t instanceof Error?t:new Error(t),null),this.callbacks.delete(e))}}]),TrustWeb3Provider}();window.Trust=TrustWeb3Provider,window.Web3=_web.default;
+},{"web3":"kdoM","./rpc":"afne","./error":"MNFS","./utils":"FOZT","./id_mapping":"sD6q","events":"wIHY"}]},{},["Focm"], null)
 //# sourceMappingURL=/trust-min.js.map
