+++ src/error.js
@@ -1,0 +1,21 @@
+// Copyright © 2017-2020 Trust Wallet.
+//
+// This file is part of Trust. The full Trust copyright notice, including
+// terms governing use, modification, and redistribution, is contained in the
+// file LICENSE at the root of the source code distribution tree.
+
+"use strict";
+
+class ProviderRpcError extends Error {
+  constructor(code, message) {
+    super();
+    this.code = code;
+    this.message = message;
+  }
+  
+  toString() {
+    return `${this.message} (${this.code})`;
+  }
+}
+
+module.exports = ProviderRpcError;
