+++ src/rpc.js
@@ -1,3 +1,9 @@
+// Copyright © 2017-2020 Trust Wallet.
+//
+// This file is part of Trust. The full Trust copyright notice, including
+// terms governing use, modification, and redistribution, is contained in the
+// file LICENSE at the root of the source code distribution tree.
+
 "use strict";
 
 class RPCServer {
@@ -20,7 +26,6 @@ class RPCServer {
   }
 
   call(payload) {
-    // console.log("==> call rpc ", payload);
     return fetch(this.rpcUrl, {
       method: "POST",
       headers: {
@@ -31,7 +36,6 @@ class RPCServer {
     })
     .then(response => response.json())
     .then(json => {
-      // console.log("<== rpc result", json);
       if (!json.result && json.error) {
         console.log("<== rpc error", json.error);
         throw new Error(json.error.message || "rpc error");
