+++ src/tests/test.js
@@ -1,3 +1,9 @@
+// Copyright © 2017-2020 Trust Wallet.
+//
+// This file is part of Trust. The full Trust copyright notice, including
+// terms governing use, modification, and redistribution, is contained in the
+// file LICENSE at the root of the source code distribution tree.
+
 "use strict";
 
 require("../index");
@@ -72,7 +78,7 @@ describe("TrustWeb3Provider constructor tests", () => {
     provider.request(request).then((chainId) => {
       expect(chainId).toEqual("0x3");
       done();
-    })
+    });
 
     const response = web3.currentProvider.send(request);
     expect(response.result).toBe("0x3");
@@ -90,7 +96,7 @@ describe("TrustWeb3Provider constructor tests", () => {
     provider.request({method: "eth_accounts"}).then((accounts) => {
       expect(accounts).toEqual(["0x5ee066cc1250e367423ed4bad3b073241612811f"]);
       done();
-    })
+    });
 
     web3.currentProvider.sendAsync({method: "eth_accounts"}, (error, data) => {
       expect(data.result).toEqual(["0x5ee066cc1250e367423ed4bad3b073241612811f"]);
